[WIP][FEATURE] Introduce PSR-14 based EventDispatcher as hook replacement
The new PSR-14 standards recommendation (currently under review phase),
is providing a new OOP way to extend a framework in an interchangeable way.

There are four components to PSR-14:
- Listeners => actual code that hooks into a certain place
- Event => Any kind of object which defines what a listener can do (or just "listen")
- ListenerProvider => a flexible registration mechanism for a framework
- Dispatcher => the main logic to be used throughout the framework

Due to it's PSR-based nature, it is very close to various other EventDispatcher
systems provided by various frameworks.

For TYPO3 we decided to:
- Provide our own ListenerProvider that checks all active extensions for listeners
within EXT:my_extension/Configuration/ListenerProviders.php - similar to how
we've done that with PSR-15 middlewares. The ListenerProvider also uses
"makeInstance"
- Provider our own EventDispatcher which is set up at the VERY part of the
bootstrap process of TYPO3, and part of the PSR-11 Container.

The purpose of this change is to fully migrate away from
- Hooks
- Hooks that modify things
- Hooks which implement interfaces for hooks
- Signals/Slots from Extbase

and their special registration into one Object-Oriented way
on how to treat Events and Listeners.

The current change provides
1. A custom PackageListenerProvider
- with the mentioned registration concept
- which allows to do ordering based on DependencyOrderingService

2. A straightforward EventDispatcher implementation

3. Events for all previous UserAuthentication hooks

4. Compatibility Listeners for all previous UserAuthentication hooks, these
are usually loaded before other listeners

5. An change to the Backend module "Configuration" to browse registered listeners

The aim for this patch is
- to be 100% compliant to the spec
- to change all BackendUserAuthentication hooks as well to new Events
- to add further tests for the new functionality
- to add documentation for the new EventDispatcher system
- to show examples on how to migrate away from SignalSlot

Resolves: #?
Releases: master
Change-Id: I4ccda33fdd7bf25e75dfaaef3f6053c706dcff7c