[WIP][FEATURE] Add FLUIDPAGE cObject
This adds a new cObject "FLUIDPAGE" which has a much simpler
usage than "FLUIDTEMPLATE", because it is meant to render
a full page like this:

fluidpage = PAGE
fluidpage.typeNum = 0
fluidpage.10 = FLUIDPAGE
fluidpage.10.templateRootPaths.100 = fileadmin/templates/

Nothing else is needed, as
a) the layoutRootPaths are resolved to be meant to be under "Templates/Layouts"
b) the partialRootPaths are resolved to be meant to be under "Templates/Partials"
automatically

in addition, the objects
* site
* siteLanguage
* page(information)

are injected automatically, so this can be managed to be worked directly.

The name of the template is resolved from the Page Layout (Backend Layout)
identifier, so add a "Templates/Pages/MyPage.html" if there is a backend layou
with "my_page" added.

Why do we do this?

We've found that FLUIDTEMPLATE is often used for page rendering,
but lacks specific usages for pages to be worked directly.

All site settings and TypoScript constants are in place within settings.

This functionality is adapted to be used for further rendering of
custom Content Elements and Content Blocks which will follow in
subsequent steps.

For this reason this API is considered experimental until TYPO3 v13 LTS.

Resolves: #?
Releases: main
Change-Id: Ieb2b665945ad671fe7e7c195b3201dbc0dbd7076